\name{MBS-package}
\alias{MBS-package}
\docType{package}
\title{
Modified bagging schema R implementation
}
\description{
MBS is a package that provides an R implementation of the modified bagging schema defined by Darius Dziuda (see references). In that definition, the context of the use of the modified bagging schema is for biomarker discovery in gene expression data (has been used with microarray and RNA-seq gene expression data). However, the approach could be useful for selection of a parsimonious set of variables separating classes in other scenarios with datasets that have up to thousands of variables.
}
\author{
Jyota Snyder <jyotas@my.ccsu.edu>. Modified bagging schema implemented was defined by Darius Dziuda.
}
\references{
Dziuda DM (2010). \emph{Data mining for genomics and proteomics: Analysis of gene and protein expression data}. Hoboken, NJ: John Wiley & Sons, Inc.
}
\keyword{ package }
