% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MBS.R
\docType{methods}
\name{MBS}
\alias{MBS}
\title{Set up Modified Bagging Schema experiment with chosen settings}
\usage{
\S4method{MBS}{matrix,numeric}(dataMatrix, classes, stopP = NULL,
  stopT2 = NULL, reps = NULL, initialSelection = "random",
  priors = NULL, proportionInBag = 0.632, searchWithReplacement = TRUE,
  assessOutOfBag = TRUE, showProgress = TRUE)

\S4method{MBS}{data.frame,numeric}(dataMatrix, classes, ...)
}
\description{
Set up Modified Bagging Schema experiment with chosen settings

Set up Modified Bagging Schema experiment with chosen settings (for data.frame)
}

