% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MBS.R
\docType{methods}
\name{mbsHybridFeatureSelection}
\alias{mbsHybridFeatureSelection}
\title{Perform hybrid feature selection with MBS object}
\usage{
\S4method{mbsHybridFeatureSelection}{MBS,numeric}(object, selectedRows)
}
\description{
Perform hybrid feature selection with MBS object
}

