% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MBS.R
\docType{methods}
\name{mbsObtainBestInitial}
\alias{mbsObtainBestInitial}
\title{Obtain best initial feature for MBS iteration}
\usage{
\S4method{mbsObtainBestInitial}{MBS,numeric}(object, selectedRows)
}
\description{
Obtain best initial feature for MBS iteration
}

