% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MBS.R
\docType{methods}
\name{MBSparallel}
\alias{MBSparallel}
\title{Run a multi-core parallelized Modified Bagging Schema experiment}
\usage{
\S4method{MBSparallel}{matrix,numeric,numeric}(dataMatrix, classes,
  cores = detectCores(), stopP = NULL, stopT2 = NULL, reps = NULL,
  initialSelection = "random", priors = NULL, proportionInBag = 0.632,
  searchWithReplacement = TRUE, assessOutOfBag = TRUE,
  showProgress = TRUE)
}
\description{
Run a multi-core parallelized Modified Bagging Schema experiment
}

