% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MBS.R
\docType{methods}
\name{mbsRun}
\alias{mbsRun}
\title{Run Modified Bagging Schema iterations for MBS object}
\usage{
\S4method{mbsRun}{MBS,logical}(object, showProgress)
}
\description{
Run Modified Bagging Schema iterations for MBS object
}

