% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MBS.R
\docType{methods}
\name{summary}
\alias{summary}
\title{summary of MBS object}
\usage{
\S4method{summary}{MBS}(object, variableUsedAccuracy = 1,
  numberOfAccurateVars = 10)
}
\description{
summary of MBS object
}

